Implement options button, remove top bar dropdown
