+++ packages/sp-frontend/src/options.style.js
@@ -1,0 +1,6 @@
+
+import styled from "styled-components";
+
+export const OptionsContainer = styled.div`
+  padding: 1em;
+`;
